feat(theme-typings): reactor src file structure
