+++ tooling/cli/src/command/tokens/config.ts
