+++ tooling/cli/src/command/tokens/extract-color-schemes.ts
