+++ tooling/cli/src/command/tokens/extract-component-types.ts
