+++ tooling/cli/src/command/tokens/extract-property-paths.ts
