+++ tooling/cli/src/utils/init-cli.ts
@@ -1,7 +1,7 @@
 import checkNode from "cli-check-node"
 import welcome from "cli-welcome"
 import updateNotifier from "update-notifier"
-import * as pkgJSON from "../package.json"
+import * as pkgJSON from "../../package.json"
 
 export function initCLI() {
   checkNode(`12`)
